refactor(registering commands): auto-register commands
