+++ packages/cli/templates/support.js
@@ -1,3 +1,6 @@
 import '@dhis2/cypress-commands'
+// You can use helpers of `@dhis2/cypress-commands` directly:
+//   import { enableAutoLogin } from '@dhis2/cypress-commands'
+//   enableAutoLogin()
 
 // Add additional support functions here
