+++ packages/cypress-commands/src/constants.js
@@ -1,0 +1,1 @@
+export const LOGIN_ENDPOINT = 'dhis-web-commons-security/login.action'
